+++ src/components/ui/block.contents.tsx
@@ -1,29 +1,5 @@
 import React, {type JSX} from 'react';
-
-export interface ContentBlock {
-  type: 'paragraph' | 'heading' | 'list' | 'list-item' | 'image' | 'link' | 'quote' | 'code';
-  level?: number;
-  image?: {
-    url: string;
-    formats?: {
-      thumbnail?: { url: string };
-      small?: { url: string };
-      medium?: { url: string };
-      large?: { url: string };
-    };
-    width: number;
-    height: number;
-    name: string;
-  },
-  children: Array<{
-    type: string;
-    code?: boolean;
-    text?: string;
-    bold?: boolean;
-    url?: string;
-    children?: Array<{ text: string; type: string; }>;
-  }>;
-}
+import { type ContentBlock } from '~/types.d';
 
 interface PageContentProps {
   params: {
