--- src/pages/about/partners.mdx
@@ -1,233 +1,0 @@
----
-title: International business network
-description: Explore our network of customers, projects, and partners, offering quick solutions for your company at special rates
-image: ~/assets/images/about/calima-service-providers.png
-layout: '~/layouts/MarkdownLayout.astro'
-openGraph:
-  images:
-    - url: '/assets/images/about/calima-service-providers.png'
-      width: 900
-      heigth: 600
----
-
-As an international conglomerate, our powerful arm reaches into multiple industries
-
-### Portfolio & Clients
-
-Our growing team of founders is collaborating to offer reliable services & support with international reach
-
-## Calima Business Marketplace
-
-Connect with service providers in our network to receive reliable services & priority support
-
-Fractional CTOs, Infrastructure management, CISO, HIPPA compliance, due diligence
-
-## Morir Soñando
-
-Artist marketplace for members of our collective
-
-Powered by Markketplace
-
-[Morir Soniando Markket Storefront](https://astro.morirsoniando.com/)
-
-## RANA EHR
-
-Calima, la rana que sana, is an electronic care coordination ecosystem
-
-With Open Source & subscription services, for better patient outcomes & data interoperability
-
-[ehr.caliman.org](https://ehr.caliman.org/)
-
-## Sandoval Collab
-
-Mental health practitioners and ketamine assited therapy
-
-[Sandoval Collab Community Resources](https://www.sandovalcolab.com/community-resources)
-
-## Siembra Today
-
-Non-profit organization devoted to transforming the lives of our community by providing accessible mental health and wellness support
-
-[Siembra today homepage](https://www.siembratoday.org)
-
-## PhaseZero
-
-PhaseZero is a HIPPA compliant messaging service, with dashboards and rest APIs to enable healthcare services to go to market faster.
-
-[PhaseZero Website](https://www.phasezero.co/)
-
-## MarkketPlace
-
-We're creating an open standards ecosystem, to help creators manage with their
-storefronts, subscriptions, payments, and media assets
-
-[markket.place](https://markket.place/about/)
-
-## Lotus AI
-
-Powered by Claimpower, medical billing services & tools for medical practices
-
-[Lotus docs](https://dev.lotuscares.ai/docs)
-
-# Partners & Service providers
-
-Reliable tools and individuals to supplement any department.
-
-Chat with our team to receive warm introductions and special offers.
-
-## Forecastr - Chief Financial Office
-
-[Forecastr](/about/partners/forecastr) is an out-of-the-box finance function for your business. We help you make better decisions and succeed in fundraising.
-
-[Forecastr website](https://www.forecastr.co/)
-
-## Medplum EMR
-
-Electronic Medical Records, with APIs & growing library of examples & components.
-
-Our engineers can manage your deployment and implementation to power your healthcare products and get to market faster.
-
-- [Medplum API docs](https://www.medplum.com/docs/api)
-- [Medplum Solutions](https://www.medplum.com/solutions)
-- [Medplum Self Hosting](https://www.medplum.com/docs/self-hosting)
-
-## Markketplace
-
-Open source tools built on top of Strapi, designed to simpliy the way
-marketplaces are created and deployed.
-
-We're building a suite of marketplace components that are  easy to
- deploy, scalable and interoperable.
-
-- [Github](https://github.com/calimania/markketplace)
-- [Strapi Documentation](https://docs.strapi.io)
-
-## NimbleHQ
-
-Team of developers, designers, and product owners taking pride in building delightful products.
-
-Based in south east asia and latin america.
-
-[nimblehq](https://nimblehq.co)
-
-## Datadog
-
-Datadog is a monitoring and security platform for developers, IT operations teams and business users in the cloud age. Our engineers can monitor your implementation and apply lessons from the insights.
-
-[Datadog getting started](https://docs.datadoghq.com/getting_started/)
-
-## Baires Dev
-
-Access thousands of prevetted software engineers and product professionals. Streamlined process, book a discovery call. We’ll discuss team structure and approach, success criteria, timescale, budget, and required skill sets to see how we can help.
-
-[Baires Dev Website](https://www.bairesdev.com/)
-
-## AWS & Activate
-
-With our partner code you can receive up to 10.000 USD in additional credits to launch your MVP faster when you apply for AWS activate.
-
-Additionally we  can help you navigate AWS and get exposure inside with other AWS sales reps, other programs and experts.
-
-[AWS Activate](https://aws.amazon.com/startups?lang=en-US)
-
-## Digital Ocean
-
-For a quick launch and fast iteration, we recommend using Digital Ocean. Get a droplet online and connected to your source code in a few minutes. Domains and Let's encrypt included.
-
-Digital Ocean fosters an active technical community, organizing events, supporting open source projects and hosting tutorials and forums.
-
-[Digital Ocean for Startups](https://www.digitalocean.com/hatch)
-
-## Rackspace
-
-Managed cloud solutions, server space, company email and multiple tools to outsource your IT and network operations and support your team.
-
-[Rackspace](https://www.rackspace.com/)
-
-## INSART fintech engineering
-
-Talented sofware engineers with experience developing products for financial services
-
-[Insart](https://insart.com/)
-
-## Pygma Fintech Accelerator
-
-Pygma is the leading incubator for fintech startups from latin America, looking to create an international network of founders providing financial services for the ecosystem.
-
-[Pygma Homepage](https://www.pygma.co/)
-
-## Hoyos & Associates Legal services
-
-Legal services for startups and fund managers.
-
-[Hoyos & Associates Law Website](https://www.hoyos.law/)
-
-## Virtustant
-
-Virtual Assistants and remote professional recruiting in Latin America.
-
-[Virtustant Homepage](https://virtustant.com/)
-
-## Martha Razo
-
-Martha is a business consultant and event producer in the Chicago area.
-
-[Martha Razo website](https://www.martharazo.com/)
-
-## Alternova
-
-Talented engineers, product and business leaders experienced in healthcare products. Open source & interoperability.
-
-- [Alternova Website](https://alternova.com/)
-- [Cardinal Kit](https://cardinalkit.org/)
-
-## Techstars Startup Weekend
-
-Learn how to create a startup in 3 days of workshops with mentors, pitch to investors and get actionable feedback to create a roadmap to get to market, and join a global network of startup enthusiasts.
-
-Apply to run an event in your local community.
-
-[Techstars Startup Weekend](https://www.techstars.com/communities/startup-weekend)
-
-## Google Workspace
-
-Our referral code grants you a small discount, and we can help create and manage your organization settings.
-
-[Google workspace](https://workspace.google.com/lp/business)
-
-## MyPage.chat: Landing Page and Email for startups
-
-Enter your full name to get a personalized link for scheduling, e-mail and more.
-
-[Mypage.chat](https://mypage.chat/)
-
-## Zoho Mail
-
-Free mail with your domain name, for 5 or fewer users.
-
-[Zoho mail](https://www.zoho.com/mail/zohomail-pricing.html)
-
-## Pump
-
-SAVE ~60% ON CLOUD FOR FREE
-
-Pump uses group buying & AI to save on cloud, so startups can now save like the big tech!
-
-[Pump](https://pump.co/)
-
-## Pinneaple Women
-
-Empoderamos a jóvenes mediante mentoría y recursos para que construyan su autoestima y reconozcan su valor. Nutrimos sus sueños y cultivamos sus talentos en preparación para un futuro brillante
-
-
-[Pinneaple women](https://www.pineapplewomen.com/)
-
-## HUE Capital
-
-Hue Capital is a business media and tech platform providing digestible
-career and business content to rising industry experts and thought
-leaders.
-
-Their community directory tool destined to help you find business groups that align with your interests.
-
-[Hue Capital Kinnected Company directory](https://www.huecapital.co/kinnected/community-directory)
