--- src/pages/about/partners/forecastr.mdx
@@ -1,79 +1,0 @@
----
-title: Forcastr - Fractional Finance Office
-description: Our trusted partner, helps startups manage finances, build investor relations, and follow proven playbooks for growth
-image: ~/assets/images/partners/forecastr.png
-layout: '~/layouts/MarkdownLayout.astro'
----
-
-> essential for startups aiming to scale effectively and make informed strategic decisions
-
-## Leveraging Fractional Executives: How Forecastr Elevates Startup Success
-
-### About Forecastr
-
-Forecastr makes financial modeling easy.
-
-Using Forecastr, founders can easily create, track and manage their financial projections in preparation
-for raising capital.
-
-Each Forecastr client is assigned a dedicated financial analyst to help them turn finance into a competitive advantage.
-
-
-## Why do customers love Forecastr:
-
-- Custom financial model built in Forecastr
-- Dedicated financial expert assigned to every customer
-- Ability to easily understand, present and maintain your financial model
-- Expert guidance and support to prepare you for investors
-- Guardrails that prevent you from 'breaking' the financial model and accidentally using misinformation
-- Confidence in accurate up-to-date data from direct integrations to your other business solutions (QBO, Gusto, Stripe, etc.)
-- Professional investor reports and Board presentation slides
-
-**********************
-
-## How they work with partners
-
-- Deep subject-matter expertise in startup financial modeling and fundraising to partners as an educational value-add for founders & startup communities
-- Forecastr provides free educational webinars, workshops, pitch events, financial modeling templates, and other resources that are beneficial for founders and communities
-
-## Current Partners Include:
-
-Techstars [50+ programs], Intuit, Microsoft for Startups, Gust, Carta, Founder Institute, HubSpot, Gener8tor, Foundersuite, HubSpot, Wefunder, and many more!
-
-## Benefits to our startups
-
-- Discounted monthly fees
-- Access to exclusive features and additional resources
-- White-glove onboarding where we work with you to build a world-class, custom financial model for your company
-- Dedicated financial analyst support
-- Access to Financial Modeling Subject Matter Experts / Mentors
-- FREE Financial Modeling Curriculum for Founders
-
-## Educational Content & Workshops
-
-- FREE Financial Model Templates
-- Monthly Finance Playbook for Startups
-
-**********************
-
-Club Calima is a happy customer and partner of Forecastr
-
-Connect with our executives to have look into our models and see the product in action
-
-> A CFO grants legitimacy to your company; partnering with the right one can add significant diligence to the due diligence process.
-
-Connect with Forecastr at their events, or message us for an introduction and special requests
-
-[Forecastr Pricing](https://www.forecastr.co/pricing)
-
-## Forecastr Free Webinar Series Events
-
-**July 11** - Keeping Investors Happy & Writing Checks - [Register Here](https://hubs.li/Q02Cph_30)
-
-**July 17** - Preparing for Fundraising: Technical Due Diligence - [Register here](https://hubs.li/Q02DmtSZ0)
-
-![Forecastr & Club Calima - Fractional teams to complete your startup](~/assets/images/partners/forecastr.png)
-
-> We put the fractional in CFO
-
--- Chief Friendship Officer
