+++ src/types.d.ts
@@ -337,3 +337,62 @@ export interface Content extends Headline, Widget {
 }
 
 export interface Contact extends Headline, Form, Widget {}
+
+
+interface BlockText {
+  text: string;
+  type?: 'text';
+  bold?: boolean;
+}
+
+interface BlockLink {
+  type: 'link';
+  url: string;
+  children: BlockText[];
+}
+
+type BlockChild = BlockText | BlockLink;
+
+export interface ContentBlock {
+  type: 'paragraph' | 'heading' | 'list' | 'list-item' | 'image' | 'link' | 'quote' | 'code';
+  type: string;
+  level?: number;
+  image?: {
+    url: string;
+    formats?: {
+      thumbnail?: { url: string };
+      small?: { url: string };
+      medium?: { url: string };
+      large?: { url: string };
+    };
+    width: number;
+    height: number;
+    name: string;
+  },
+  children: Array<{
+    type: string;
+    code?: boolean;
+    text?: string;
+    bold?: boolean;
+    url?: string;
+    children?: Array<{ text: string; type: string; }>;
+  }>;
+}
+
+export interface Page {
+  id: number;
+  Title: string;
+  Content: ContentBlock[];
+  Active: boolean;
+  slug: string;
+  createdAt: string;
+  updatedAt: string;
+  publishedAt: string;
+  locale: string;
+  menuOrder: number | null;
+  documentId: string;
+  store?: Store;
+  SEO?: SEO;
+  albums?: Album[];
+}
+
